# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Seatunnel X",
        "contact": {},
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/oauth/callback": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "oauth"
                ],
                "parameters": [
                    {
                        "description": "request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/oauth.CallbackRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/oauth.CallbackResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/oauth/login": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "oauth"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "OAuth provider (github, google)",
                        "name": "provider",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/oauth.GetLoginURLResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "oauth.CallbackRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
        "oauth.CallbackResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error_msg": {
                    "type": "string"
                }
            }
        },
        "oauth.GetLoginURLResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "error_msg": {
                    "type": "string"
                }
            }
        },
...
    }
}